Create data-analysis
An analysis of the raw collected data to calculate the U-value (thermal efficiency) of a room.
Exponential curves are fitted to decays in CO2 and temperature which overlap to find the decay coefficients of both elements. Then, a Monte Carlo analysis allows propagation of uncertainty to be diminished and for a final, accurate result to be found.